restructured board component but rotation of cards while dragging is still not working
