+++ src/app/main-content/board/board.component.scss
@@ -131,36 +131,54 @@ section {
     }
 }
 
-.card {
-    width: 252px;
-    min-width: 252px;
-    background-color: #FFFFFF;
-    border-radius: 24px;
-    padding: 16px;
-    margin-bottom: 16px;
+// .cdk-drag {
+//     transform: rotate(-5deg);
+// }
 
-    &:hover {
-        cursor: pointer;
-        scale: 1.02;
-    }
-
-    // &.cdk-drag-preview {
-    //     transform: rotate(-5deg);
-    // }
+// .cdk-drag-preview {
+//     transform: rotate(-5deg);
+// }
     
-    &.cdk-drag-placeholder {
-        opacity: 0.5;
-        border: 1px dashed #A8A8A8;
-    }
+// .cdk-drag-placeholder {
+//     opacity: 0.5;
+//     border: 1px dashed #A8A8A8;
+// }
 
-    // &.cdk-drag-dragging:active {
-    //     animation: rotating 0.4s linear infinite; // Fügt eine Dreh-Animation hinzu
-    // }
+// .card_wrapper {
+//     width: 252px;
+//     min-width: 252px;
+// }
 
-    &.dragging {
-        transform: rotate(-5deg);
-    }
-}
+// .card {
+//     width: 252px;
+//     min-width: 252px;
+//     background-color: #FFFFFF;
+//     border-radius: 24px;
+//     padding: 16px;
+//     margin-bottom: 16px;
+
+//     &:hover {
+//         cursor: pointer;
+//         scale: 1.02;
+//     }
+
+//     // &.cdk-drag-preview {
+//     //     transform: rotate(-5deg);
+//     // }
+    
+//     &.cdk-drag-placeholder {
+//         opacity: 0.5;
+//         border: 1px dashed #A8A8A8;
+//     }
+
+//     // &.cdk-drag-dragging:active {
+//     //     animation: rotating 0.4s linear infinite; // Fügt eine Dreh-Animation hinzu
+//     // }
+
+//     &.dragging {
+//         transform: rotate(-5deg);
+//     }
+// }
 
 // @keyframes rotating {
 //     0% { transform: rotate(0deg); }
@@ -180,16 +198,20 @@ section {
     justify-content: center;
 }
 
+section .custom-dialog {
+    padding: 0;
+}
 // open card dialog
 .custom-dialog {
-    position: fixed;
-    top: 0;
-    right: 0;
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
+    // position: fixed;
+    // top: 0;
+    // right: 0;
+    // width: 100%;
+    // height: 100vh;
+    // display: flex;
+    // justify-content: center;
+    // align-items: flex-start;
+    padding: 0;
     z-index: 100;
     animation: slideInFromRight 125ms ease-out forwards;
 }
@@ -218,7 +240,7 @@ section {
 }
 
 .dialog-content {
-    width: 950px;
+    width: 890px;
     height: 700px;
     padding: 20px;
     background-color: #fff;
