+++ src/app/main-content/board/card/card.component.ts
@@ -57,7 +57,7 @@ export class CardComponent {
     this.isDialogOpen = true;
   }
   
-   closeEditDialog(){
+  closeEditDialog(){
     this.isDialogOpen = false; 
   }
