+++ src/app/main-content/board/edit-dialog/edit-dialog.component.ts
@@ -24,6 +24,7 @@ export class EditDialogComponent {
   subtaskInput: string = '';
   subtasks: { name: string, isEditing: boolean }[] = []; // Array für Subtasks
   dropdownVisible: boolean = false;
+  isEditFormSubmitted: boolean = false;
   newDate: string = "";
 
   assignContact(contactId: string) {
@@ -63,7 +64,8 @@ export class EditDialogComponent {
   }
 
   saveEditedTask(taskForm: NgForm) {
-    if (this.item?.id && this.item?.date) {
+    this.isEditFormSubmitted = true;
+    if (this.item?.id && this.item?.date && this.isEditFormSubmitted) {
       this.firebase.updateTaskStatus(this.item.id, {
         title: this.item.title,
         description: this.item.description,
