+++ src/app/main-content/board/task-card/task-card.component.html
@@ -1,39 +1,49 @@
-<span class="category" [ngStyle]="{'background-color': item.category === 'Technical Task' ? '#1FD7C1' : '#0038FF'}">{{item.category}}</span>
-<p class="title">{{item.title}}</p>
-<p class="description">{{item.description}}</p>
-<div class="subtasks">
-    @if(item.subtasks) {
-        @if (item.subtasks.length !== 0) {
-            <div class="progress_bar">
-                <div class="progress" [ngStyle]="{'width': getCompletedPercentage() + '%'}"></div>
-            </div>
-            <p>{{completedSubtasks}}/{{item.subtasks.length}} Subtasks</p>
-        }
-    }
+<div *ngIf="isDialogOpen" class="overlay_content" (click)="closeDialog()">
+    <div class="custom-dialog">
+        <app-card [item]="selectedItem" (closeDialogEvent)="closeDialog()" (click)="stopPropagation($event)"></app-card>
+    </div>
 </div>
-<div class="content_wrapper">
-    <div class="assigned_to">
-        @if(item.assignedToUserId && item.assignedToUserId.length > 0) {
-            @for (userId of item.assignedToUserId; track $index) {
-                <div class="avatar_wrapper">
-                    @for (contact of firebase.orderedContactsList; track $index) {
-                        @if(contact.id === userId) {
-                            <div class="avatar" [ngStyle]="{'background-color': contact.color}">
-                                <p>{{contact.firstname.charAt(0)}}{{contact.lastname.charAt(0)}}</p>
-                            </div>
-                        }
-                    }
+<div class="card" cdkDrag (click)="openCardDialog(item)"
+    (cdkDragStarted)="item.id ? onDragStart(item) : null"
+    (cdkDragEnded)="item.id ? onDragEnd(item) : null"
+    [ngClass]="{ 'dragging': item.id ? isCardDragging(item.id) : false }">
+    <span class="category" [ngStyle]="{'background-color': item.category === 'Technical Task' ? '#1FD7C1' : '#0038FF'}">{{item.category}}</span>
+    <p class="title">{{item.title}}</p>
+    <p class="description">{{item.description}}</p>
+    <div class="subtasks">
+        @if(item.subtasks) {
+            @if (item.subtasks.length !== 0) {
+                <div class="progress_bar">
+                    <div class="progress" [ngStyle]="{'width': getCompletedPercentage() + '%'}"></div>
                 </div>
-            } 
+                <p>{{completedSubtasks}}/{{item.subtasks.length}} Subtasks</p>
+            }
         }
     </div>
-    <div class="priority">
-        @if (item.priority === "urgent") {
-            <img src="./assets/tasks/Prio alta.svg" alt="">
-        } @else if(item.priority === "medium") {
-            <img src="./assets/tasks/Prio media.svg" alt="">
-        } @else if(item.priority === "low") {
-            <img src="./assets/tasks/Prio baja.svg" alt="">
-        }
+    <div class="content_wrapper">
+        <div class="assigned_to">
+            @if(item.assignedToUserId && item.assignedToUserId.length > 0) {
+                @for (userId of item.assignedToUserId; track $index) {
+                    <div class="avatar_wrapper">
+                        @for (contact of firebase.orderedContactsList; track $index) {
+                            @if(contact.id === userId) {
+                                <div class="avatar" [ngStyle]="{'background-color': contact.color}">
+                                    <p>{{contact.firstname.charAt(0)}}{{contact.lastname.charAt(0)}}</p>
+                                </div>
+                            }
+                        }
+                    </div>
+                } 
+            }
+        </div>
+        <div class="priority">
+            @if (item.priority === "urgent") {
+                <img src="./assets/tasks/Prio alta.svg" alt="">
+            } @else if(item.priority === "medium") {
+                <img src="./assets/tasks/Prio media.svg" alt="">
+            } @else if(item.priority === "low") {
+                <img src="./assets/tasks/Prio baja.svg" alt="">
+            }
+        </div>
     </div>
 </div>
