+++ src/app/main-content/board/task-card/task-card.component.scss
@@ -1,3 +1,62 @@
+.custom-dialog {
+    position: fixed;
+    top: 0;
+    right: 0;
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 100;
+    animation: slideInFromRight 125ms ease-out forwards;
+}
+
+.overlay_content {
+    background-color: rgba(0, 0, 0, 0.349);
+    position: fixed;
+    top: 0;
+    right: 0;
+    width: 100%;
+    height: 100vh;
+    z-index: 1000;
+}
+
+.card {
+    width: 252px;
+    min-width: 252px;
+    background-color: #FFFFFF;
+    border-radius: 24px;
+    padding: 16px;
+    margin-bottom: 16px;
+
+    &:hover {
+        cursor: pointer;
+        scale: 1.02;
+    }
+
+
+
+    // &.cdk-drag-dragging:active {
+    //     animation: rotating 0.4s linear infinite; // Fügt eine Dreh-Animation hinzu
+    // }
+
+    // &.dragging {
+    //     transform: rotate(-5deg);
+    // }
+}
+
+.cdk-drag {
+    transform: rotate(-5deg);
+}
+
+.cdk-drag-preview {
+    transform: rotate(-5deg);
+}
+
+.cdk-drag-placeholder {
+    opacity: 0.5;
+    border: 1px dashed #A8A8A8;
+}
 
 .category {
     color: white;
